refactor: tighten webhook header forwarding internals
